tablet: Fix sending delta for relative dials
delta takes int, not wl_fixed


(cherry picked from commit 9e98f25869a82d15db01c9c02c6843283631a462)

+++ src/wayland/tablet_v2.h
@@ -238,7 +238,7 @@ class KWIN_EXPORT TabletPadDialV2Interface : public QObject
 public:
     virtual ~TabletPadDialV2Interface();
 
-    void sendDelta(qreal delta);
+    void sendDelta(qint32 delta);
     void sendFrame(quint32 time);
 
 private:
